requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Firebase
firebase-admin==6.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development
python-multipart==0.0.6
